Change some messages and scores directory path
success message is not different when used on message dialog and on
log output, to allow for one to use html tags for formatting purposes
and the other be free of such tags.

scores path in /var/run/Minesweeper gives some problems, as some
linux installations (mainly ubuntu) mounts /var/run as ram disk, so
contents are eliminated between reboots.

Signed-off-by: Luis Colorado <luis.colorado@spindrive.fi>